hotfix: fixes crash on uncompilable graphql query
GraphQL chokes now that there are no project images with a `hero` key. I
used the dummyImages hack, this time adding the `hero` key. I updated
the projects display to check for dummy images.

Ultimately we should be able to remove the list of images on projects
entirely - I just want to make sure this is desired first.